refactor: file-browser component
